Split plugin resource packaging helpers
Keep wrac_xtask build command files below the repository quality limit while preserving the staged resource packaging behavior added for plugin resources.
